v1.0.20: the kaizo lane's engine fixes come home, and the music streams
Six edits made to the VENDORED engine in kaizo-knight-sim under its law-6
escape hatch ("port it back the same day") were still pending in its
docs/VENDOR.md. Each was proven both ways there; here the receipt is the
same one: the six whole-fight traces regenerated byte-identical, 60/60
green. Copied verbatim from that repo's copies (line endings normalised);
every site carries its own measurement in the comment.

  * sim/rng.js gmlRandomRange normalises its argument order — GML returns
    min(a,b) + f*|b-a|, and a call whose arguments invert at runtime
    (quickslash's `random_range(-spawn_yrange, spawn_yrange)` with the -1
    sentinel) got the exact negation. _tok3 f4515: y 168.674 -> 169.983.
  * sim/masks.js: a RotatedRect sprite stores NO bitmap, so the two
    diamond-blade masks are an oriented-box test (maskHitsRotatedRect, A's
    set pixels as 1x1 cells) instead of a synthesised one-row grid with
    sub-pixel holes (the recording's own blade at 311.5 sat in one); and
    masksOverlapRectA walks A on its own ink lattice at its raw position —
    the old integer-grid bound lost A's last ink column at a fractional
    position (f8088, a tooth against the mid-frame soul at 107.7478). Both
    are bit-identical for every integer-A case, which is every calibration
    dataset.
  * sim/index.js: the gravity recomposition narrows atan2 to f32 before the
    degree conversion (f6604: 197.0186767578 exact; f64 was one ulp high).
  * sim/tension.js stepGraze: a replayed pairing table is processed in the
    feed's row order (the awards are not independent — a burst paid first can
    silence a same-frame trickle), and a feed that runs out hands back to the
    geometric test instead of awarding nothing. Replay path only; free play
    has no table.
  * render/audio.js: MUSIC STREAMS, EFFECTS DECODE. decodeAudioData expanded
    the 2 MB track to raw PCM and stalled the frame it was cued on; an
    <audio> element through createMediaElementSource starts on the first
    buffered chunk and sits in the same gain graph, so the slider, the
    ceiling and stopLoop are unchanged. Falls back to the decode path if the
    browser refuses. Plus `createAudio({ overrides })`, a lane's own cue
    files by name. stopAll still forgets the loops it was waiting for.

+++ sim/index.js
@@ -185,7 +185,17 @@ function runMotion(state) {
       hs = Math.fround(hs + Math.fround(Math.fround(e.gravity) * Math.fround(gc)));
       vs = Math.fround(vs + Math.fround(Math.fround(e.gravity) * -Math.fround(gsn)));
       e.speed = Math.sqrt(hs * hs + vs * vs);
-      let dir = (Math.atan2(-vs, hs) * 180) / Math.PI;
+      // THE ATAN2 RESULT IS NARROWED BEFORE THE DEGREE CONVERSION, and the
+      // conversion itself stays f64. MEASURED on the first frame the kaizo
+      // gate's blade drift appears -- oracle f6604, a revised-tunnel blade
+      // with speed 1.400272011756897 at direction 201.81463623046875 under
+      // gravity 0.4 at 180, giving hs -1.6999996900558472 and
+      // vs 0.5203483700752258. The recording's next direction is
+      // 197.0186767578; f64 atan2 all the way gives 197.0186920166 (one f32
+      // ulp high) and narrowing the atan2 first gives 197.0186767578 exactly.
+      // Dividing by a single-precision pi instead does NOT reproduce it, so
+      // the narrowing is on the angle, not on the constant.
+      let dir = (Math.fround(Math.atan2(-vs, hs)) * 180) / Math.PI;
       if (dir < 0) dir += 360;
       e.direction = dir;
     }
